Another step on the normalization of the code
Now every element *must* have an associated renderer, that is, a renderer that the actual
rendering pipe should use to draw things, each element is actual a renderer too but we must
not use that as the renderer passed to the gfx implementation of an element. For the widget
case that's simple, the associated renderer is the theme, so everything works ok. For the
wrapper the associated element should be the wrapped element. Still need to fix the layout case
but the logic makes sense now.